修复打开图片界面置顶按钮不可见的BUG
