+++ manifest.json
@@ -4,7 +4,7 @@
     "name": "窗口置顶",
     "slug": "window_on_top",
     "description": "添加窗口置顶按钮",
-    "version": "0.1.3",
+    "version": "0.1.4",
     "author": [
         {
             "name": "沫烬染",
@@ -12,8 +12,8 @@
         }
     ],
     "injects": {
-        "renderer": "./renderer.js",
-        "main": "./main.js",
-        "preload": "./preload.js"
+        "renderer": "./src/renderer.js",
+        "main": "./src/main.js",
+        "preload": "./src/preload.js"
     }
 }
