--- preload.js
@@ -1,6 +1,0 @@
-const { contextBridge, ipcRenderer } = require("electron");
-
-
-contextBridge.exposeInMainWorld("window_on_top", {
-    toggleWindowOnTop: () => ipcRenderer.invoke("betterQQNT.window_on_top.toggleWindowOnTop"),
-});
