+++ src/preload.js
@@ -1,0 +1,6 @@
+const { contextBridge, ipcRenderer } = require("electron");
+
+
+contextBridge.exposeInMainWorld("window_on_top", {
+    toggleWindowOnTop: () => ipcRenderer.invoke("LiteLoader.window_on_top.toggleWindowOnTop"),
+});
